requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
altgraph==0.17.4
ane-transformers==0.1.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
av==12.1.0
Babel==2.15.0
bangla==0.0.2
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.34.133
botocore==1.34.133
cached_path==1.6.3
cachetools==5.3.3
catalogue==2.0.10
cattrs==23.2.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cn2an==0.5.22
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
coremltools==7.2
ctranslate2==4.3.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
Distance==0.1.3
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
dtw-python==1.5.1
einops==0.8.0
email_validator==2.2.0
encodec==0.1.1
eng-to-ipa==0.0.2
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.111.0
fastapi-cli==0.0.4
faster-whisper==1.0.2
ffmpy==0.3.2
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
fugashi==1.3.2
g2p-en==2.1.0
g2pkk==0.1.2
gast==0.5.4
gguf==0.6.0
glcontext==2.5.0
glfw==2.7.0
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
gradio==4.37.1
gradio_client==1.0.2
greenlet==3.2.1
grpcio==1.64.0
gruut==2.3.4
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
h5py==3.11.0
hangul-romanize==0.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_resources==6.4.0
inflect==7.0.0
install==1.3.5
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jaconv==0.3.4
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.3.3
kiwisolver==1.4.5
langcodes==3.4.0
langid==1.1.6
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.9.1
llama_cpp_python==0.2.26+cu118
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mecab==0.996.3
mecab-python3==1.0.9
mkl==2021.4.0
ml-dtypes==0.3.2
moderngl==5.10.0
moderngl-window==2.4.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multipledispatch==1.0.0
murmurhash==1.0.10
namex==0.0.8
nest-asyncio==1.6.0
networkx==2.8
nltk==3.8.1
num2words==0.5.13
numba==0.60.0
numpy==1.24.3
ollama==0.2.0
onnxruntime==1.18.0
openai==1.30.5
openai-whisper==20231117
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pefile==2024.8.26
pgvector==0.4.1
pillow==10.3.0
pipdeptree==2.23.0
platformdirs==4.2.2
pocketsphinx==5.0.3
pooch==1.8.1
preshed==3.0.9
proces==0.1.7
prompt_toolkit==3.0.45
proto-plus==1.24.0
protobuf==3.20.3
psutil==5.9.8
psycopg2-binary==2.9.10
pure-eval==0.2.2
pvporcupine==3.0.2
pyaml==24.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
pygame==2.5.2
pyglet==2.0.15
pyglfw==0.2.2
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pykakasi==2.2.1
pynndescent==0.5.12
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.1.2
pypinyin==0.50.0
pyreadline3==3.4.1
pyrr==0.10.3
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-weather==2.0.3
pytz==2024.1
pywhispercpp==1.2.0
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
resampy==0.4.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.10
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.0
scikit_build_core==0.9.7
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.40
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12.1
tbb==2021.13.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tornado==6.4
tqdm==4.66.4
trainer==0.0.36
traitlets==5.14.3
transformers==4.41.2
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
umap-learn==0.5.6
Unidecode==1.3.7
unidic-lite==1.0.8
urllib3==2.2.2
uvicorn==0.30.1
wasabi==1.1.3
watchfiles==0.22.0
wavmark==0.0.3
wcwidth==0.2.13
weasel==0.4.1
websockets==11.0.3
Werkzeug==3.0.3
whisper-timestamped==1.14.2
winshell==0.6
wrapt==1.16.0
yarl==1.9.4
